Add MatrixTestByJunit
