+++ java/numeric/Fraction.java
@@ -1,3 +1,5 @@
+
+
 package numeric;
 
 public class Fraction implements Comparable<Fraction> {
